+++ src/app.module.ts
@@ -5,11 +5,13 @@ import { Module } from '@nestjs/common';
 import { AuthModule } from './auth/auth.module';
 //import { UsersModule } from './users/users.module';
 import { FileModule } from './file/file.module';
+import { ReviewController } from './review/review.controller';
+import { ReviewModule } from './review/review.module';
 
 @Module({
   //imports: [AuthModule, UsersModule],
   //controllers: [AuthController],
-  imports: [AuthModule, FileModule], 
+  imports: [AuthModule, ReviewModule],
   //providers: [AppService],
 })
 export class AppModule {}
