+++ src/auth/auth.controller.ts
@@ -2,7 +2,6 @@ import { Controller, Request, UseGuards, Post, Get, StreamableFile, Res } from '
 import { LocalAuthGuard } from './local-auth.guard';
 import { AuthService } from './auth.service';
 import { JwtAuthGuard } from './jwt-auth.guard';
-import { FileService } from 'src/file/file.service';
 import { createReadStream } from 'fs';
 import { join } from 'path';
 import type { Response } from 'express';
@@ -10,8 +9,7 @@ import type { Response } from 'express';
 @Controller('auth')
 export class AuthController {
     constructor(
-      private authService: AuthService,
-      private fileService: FileService
+      private authService: AuthService
       ) {}
     @UseGuards(LocalAuthGuard)
     @Post('login')
@@ -20,16 +18,17 @@ export class AuthController {
     }
 
 
-    @UseGuards(JwtAuthGuard)
-    @Get('file')
-    async getFile() {
-      const csvFilePath = 'appstore_544007664.csv';
-      const csv = require('csvtojson');
-      csv()
-      .fromFile(csvFilePath)
-      .then((jsonObj)=>{
-          console.log(jsonObj);
-    })
-    const jsonArray=await csv().fromFile(csvFilePath);
-  }
+  //   @UseGuards(JwtAuthGuard)
+  //   @Get('file')
+  //   async getFile() {
+  //     const csvFilePath = 'appstore_544007664.csv';
+  //     const csv = require('csvtojson');
+  //     csv()
+  //     .fromFile(csvFilePath)
+  //     .then((jsonObj)=>{
+  //         console.log(jsonObj);
+  //   })
+  //   const jsonArray=await csv().fromFile(csvFilePath);
+  //   console.log(jsonArray)
+  // }
 }
