+++ src/auth/auth.module.ts
@@ -20,7 +20,7 @@ import { FileService } from 'src/file/file.service';
     }),
   ],
   controllers: [AuthController],
-  providers: [AuthService, LocalStrategy, JwtStrategy, FileService],
+  providers: [AuthService, LocalStrategy, JwtStrategy],
   exports: [AuthService]
 })
 export class AuthModule {}
