+++ src/review/review.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ReviewController } from './review.controller';
+import { ReviewService } from './review.service';
+
+@Module({
+    controllers: [ReviewController],
+    providers: [ReviewService]
+})
+export class ReviewModule {}
